Refactor file locations, layout update
