+++ styles/components/_misc.scss
@@ -95,8 +95,8 @@
 }
 
 .emoji {
-  height: 1em;
-  width: 1em;
+  height: 1em !important; // don't allow parent text sizes to override this
+  width: 1em !important;  // don't allow parent text sizes to override this
 
   &.fade {
     opacity: .35;
