+++ styles/modules/_userPage.scss
@@ -189,6 +189,10 @@
 
     .userPageReputation {
 
+      .repBg {
+        font-size: 38px;
+      }
+
       .statsBox {
         min-height: 150px;
       }
